+++ ls_UI/core/changelog.lua
@@ -1,0 +1,14 @@
+local _, ns = ...
+local E, C, PrC, M, L, P = ns.E, ns.C, ns.PrC, ns.M, ns.L, ns.P
+
+-- Lua
+local _G = getfenv(0)
+
+-- Mine
+E.CHANGELOG = [[
+- Added 9.2.7 support;
+- Added new "Info" panel to the config. It has links to various resourced including my new Discord server dedicated to
+  LS: * addons.
+- Updated German translation. Translated by nicobert90@Curse;
+- Updated embeds.
+]]